{"schema":"decodex/commit/1","summary":"Modularize live overlay layer classes","authority":"manual"}
